Feat: factors v0.1
This version is able to:
- Factorize prime number
- Factorize even number
- Factorize odd number

Is not able to:
- Factorize large numbers

Files affected;
	modified-executable:   factors
	modified:   main.c
	modified:   rsa.c
	modified:   rsa.h

Usage:
- ./factors <file>

+++ rsa.h
@@ -5,8 +5,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-int is_prime(unsigned int);
-int is_even(unsigned int);
-void find_factors(unsigned int);
+int is_prime(long int);
+int is_even(long int);
+void find_factors(long int);
 
 #endif
